fixed bug with delete cardlet

Approved-by: Kyryll Kravtsov <kyryll.kravtsov@gmail.com>

+++ src/main/java/itomy/sigterra/service/CardletViewService.java
@@ -11,6 +11,7 @@ import itomy.sigterra.service.mapper.CardletLinksMapper;
 import itomy.sigterra.web.rest.errors.CardletNotFound;
 import itomy.sigterra.web.rest.errors.CustomParameterizedException;
 import itomy.sigterra.web.rest.vm.*;
+import org.apache.commons.lang3.StringUtils;
 import org.json.JSONException;
 import org.json.JSONObject;
 import org.slf4j.Logger;
@@ -351,8 +352,14 @@ public class CardletViewService {
      */
     public void delete(Cardlet cardlet) {
         if (cardlet.getCardletHeader() != null) {
-            deleteFile(cardlet.getCardletHeader().getLogo());
-            deleteFile(cardlet.getCardletHeader().getPhoto());
+            String url = cardlet.getCardletHeader().getLogo();
+            if (StringUtils.isNotBlank(url)) {
+                deleteFile(url);
+            }
+            url = cardlet.getCardletHeader().getLogo();
+            if (StringUtils.isNotBlank(url)) {
+                deleteFile(url);
+            }
         }
     }
 
@@ -362,7 +369,14 @@ public class CardletViewService {
      * @param path
      */
     private void deleteFile(String path) {
-        awss3BucketService.deleteFile(path);
+        if (StringUtils.isNotBlank(path)) {
+            try {
+                awss3BucketService.deleteFile(path);
+                log.info("Deleted file: "+path);
+            } catch (Exception e){
+                log.error("Error delete file " + path,e);
+            }
+        }
     }
 
     /**
